requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
psycopg==3.1.18
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
django-import-export==3.0.2
psycopg==3.1.18
psycopg2-binary==2.9.9
ruff==0.1.5
black>23.0.0
pre-commit==3.7.0